Bug #316 - add platform-consistent random number generator

+++ src/tcpedit/tcpedit_types.h
@@ -147,7 +147,7 @@ typedef struct {
     tcpr_cidrmap_t *dstipmap;
     
     /* pseudo-randomize IP addresses using a seed */
-    int seed;
+    uint32_t seed;
     
     /* rewrite tcp/udp ports */
     tcpedit_portmap_t *portmap;
@@ -156,7 +156,7 @@ typedef struct {
     bool mtu_truncate;       /* Should we truncate frames > MTU? */
     int maxpacket;          /* L2 header + MTU */
 
-    unsigned long fuzz_seed;
+    uint32_t fuzz_seed;
 } tcpedit_t;
